SmokeTrail(s) moved to Fluff, variables' names fixed, floats rounded
